sync

+++ Board.cpp
@@ -1,14 +1,12 @@
 ﻿#include "pch.h"
+#include "Board.h"
 
 #include <future>
 #include <algorithm>
+#include <random>
+#include <iostream>
 
 #include "gsl/include/gsl"
-
-#include "Board.h"
-
-// for visualization purposes (0,0) is the top left.
-// as x increases move right, as y increases move down
 
 // optimized to never use std::endl until the full board is done printing
 std::wostream& operator<<(std::wostream& stream, Board& board)
@@ -37,7 +35,7 @@ void Board::Resize(uint16_t width, uint16_t height, uint16_t maxage)
 	_height = height;
 	_width = width;
 	_maxage = maxage;
-	_board.resize(_height * _width);
+	_board.resize(gsl::narrow_cast<size_t>(_height * _width));
 
 	SetThreadCount();
 }
